// src/app/models/masters/bank.ts
export class Bank {
    id: number=-1;
    name: string;
    bankTypeId: number;
    countryId: number;
    bankType: string;
    country: string;
    accountCode: string;
    swiftCode: string;
    sudebanCode: string;
    active: boolean;
    binaryImage: string;
    image: string;
    createdByUserId: number;
    accountingAccountCode:string=""
    accountingAccountId:number=-1
    accountingAccountName:string=""
    indPermiteAuxiliar:boolean=false

    
}